STU 2018 D & F
D was pretty simple, simply list all the possible die configurations. Made a slightly cleaner solution with rotation checking just to practice, but since there are so few possibilities, it was easier to enumerate them all.

F was a lot more difficult. The part that was relatively easy was finding if given point A, what are the points B,C,D, such that the points form a square (points are integer coordinates, but square is not parallel to the axes). The key to finding the correct solution is realizing that, because the points given  must make up a valid solution, there cannot be an overlap, nor a smaller square completely contained within a larger square, unless the larger square is unnecessary. This means that as long as you consider the point A in some predetermined order (-x,-y to +x,+y), and you make the smallest table possible each time, the end configuration will be correct.